Move "/filter-by-attribute-bounds" methods to data_objects
